Conflicts:
	libraries/GCS_MAVLink/include/mavlink/v1.0/ardupilotmega/ardupilotmega.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/ardupilotmega/mavlink_msg_mount_status.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/ardupilotmega/version.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/common/common.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/common/mavlink_msg_hil_optical_flow.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/common/mavlink_msg_ping.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/common/testsuite.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/common/version.h
	libraries/GCS_MAVLink/include/mavlink/v1.0/mavlink_types.h

+++ libraries/AP_InertialSensor/AP_InertialSensor_L3GD20.cpp
@@ -453,7 +453,7 @@ void AP_InertialSensor_L3GD20::disable_i2c(void)
 	hal.scheduler->panic(PSTR("L3GD20: Unable to disable I2C"));
 }
 
-uint8_t AP_InertialSensor_L3GD20::set_samplerate(uint8_t frequency)
+uint8_t AP_InertialSensor_L3GD20::set_samplerate(uint16_t frequency)
 {
 	uint8_t bits = REG1_POWER_NORMAL | REG1_Z_ENABLE | REG1_Y_ENABLE | REG1_X_ENABLE;
 	if (frequency == 0)
